requirements.txt
streamlit
pandas
numpy
streamlit-aggrid
plotly
python-dotenv
requests